Renaming dialogue handlers for consistency.
I had intended to rename the 'save' dialogue handle to 'viewer_save' but that
turned out to clash with the existing viewer_save function. Instead perform a
more extensive renaming of the save, scale and info dialogue handling
functions: suffixing their names with 'dlg'.

It's not strictly correct as they're not all dialogues but instead the
viewer's code for handling dialogues, but it will do for now. Also, I've not
renamed the files to match.


git-svn-id: svn://svn.riscos.info/privateeye/trunk@51 befe153f-808b-4961-b90b-b6e528e3d82e