Refactored the histogram and metadata gadget windows to remove shared code.
Created an image observing window library (imageobwin) to hold the common
code. This is similar to the dialogue code in that it provides a struct which
is installed as the first member of the user's struct. A __construct function
is then called to set up the object, passing in a series of callback
functions including wimp event handlers.

This struct provides a 'factory' which manufactures new image observing
objects - imageobwin_t - when requested using imageobwin_open.

Also:

- imageobserver has had an opaque handle pointer added. The previous code
  never needed to rely on it for context passing, but the new code does.

- 'hist' message tokens have been renamed 'histogram' so that the imageobwin
  code can be passed a single base token from which it can work out other
  token names.

- Histogram and metadata configuration parameters have been marshalled into
  their own structures so they can be passed around via opaque pointers. This
  is internal presently but I may move it out to the respective interfaces.

- Image observer windows are now closed if they are open but their available
  function changes state such that they're no longer valid, i.e. if a JPEG is
  converted into a sprite and the metadata vanishes, then it's no longer
  valid to keep the window open.



git-svn-id: svn://svn.riscos.info/privateeye/trunk@58 befe153f-808b-4961-b90b-b6e528e3d82e