Fixes to FFG conversion system.
The Wimp message references received from the translators I've been testing
against seem to be nonsensical/corrupt values, so instead of just tracking
the message reference numbers I need to track the task handle of the
converter. (If the translator persisted in memory and handled multiple
conversions then this could go wrong.)

The translator I've been testing also fails to set the file type of the
converted file correctly, so ffg_complete also forces the file type to the
expected value.

- Added ffg_apposite which lets the dataxfer code know if an FFG translation
  is taking place.
- Added ffg_complete which resets the stored translator task handle at the
  end of the session.
- Added a message handler for message_TRANSLATE_FFG_ACK. We use this to
  remember the translator's task handle.
- messageack_translate_ffg: Don't check the task handle is ours, because it
  ought to be ours received in the bounced message.

Also:

- Now uses linked list library to manage the list of known translators.
- ffg__loadable typedef added.
- return event_HANDLED; instead of return 1; in message handlers.


git-svn-id: svn://svn.riscos.info/privateeye/trunk@148 befe153f-808b-4961-b90b-b6e528e3d82e