Improve tag cloud layout.
Previously we would give the every line the same amount of space, which we
call 'leading'. We now give the topmost line a 'font size' worth of space and
space subsequent lines out by 'font size + leading'. The final line's has
leading added to it.

The result is a visually more consistent amount of border around the text and
the removal of a fudge from redraw.c.



git-svn-id: svn://svn.riscos.info/privateeye/trunk@180 befe153f-808b-4961-b90b-b6e528e3d82e