Templates: Pass through my bug-fixed CCres
CCres would output Templates with empty validation strings encoded as a pointer to a carriage return. This seems to trigger a bug in the Wimp which causes a nearby task to be paged in, and its address space walked upon, which can cause a crash if that task has a smaller WimpSlot than us. The modified CCres fixes that encoding to output a -1 as expected. It also avoids leaving cruft inside the Template file by making efforts to wipe spare buffer bytes.